Wrote basic implemetation for ShootAction
For the ShootAction class, source and destination instance variables of type GameActor were made. This is because in whatever case a ShootAction occurs, both will either be a Player or Alien who inherit from that class. A basic init method was written that will likely be changed, as well as a basic execute method to conform to the PGameAction protocol.